refactor(ui): replace old Pod components with new KVM components
