--- ui/src/app/kvm/components/PodAggregateResources/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./PodAggregateResources";
