--- ui/src/app/kvm/components/PodResourcesCard/_index.scss
@@ -1,209 +1,0 @@
-@mixin PodResourcesCard {
-  $donut-size: 6.5rem;
-  $donut-width: 14px;
-
-  .pod-machine-list-modal {
-    max-width: none;
-    min-width: 0;
-    padding: $spv-inner--x-small $sph-inner--small;
-    // Subtract small screen gutter width and card inner padding to center dropdown.
-    width: calc(
-      100vw - #{(map-get($grid-gutter-widths, small) + $sph-inner) * 2}
-    );
-
-    .fqdn-col,
-    .status-col {
-      width: 50%;
-    }
-
-    .power-col {
-      width: 0;
-    }
-
-    .cores-col,
-    .ram-col {
-      width: 5rem;
-    }
-
-    // Overwrite overflow: hidden here where there's no checkbox and width is static.
-    td:first-child .p-double-row__primary-row-text {
-      overflow: inherit !important;
-    }
-
-    @media only screen and (min-width: $breakpoint-x-small) {
-      .power-col {
-        width: 7rem;
-      }
-    }
-
-    @media only screen and (min-width: $breakpoint-small) {
-      width: 35rem;
-    }
-
-    @media only screen and (min-width: $breakpoint-medium) {
-      width: 40rem;
-    }
-  }
-
-  .pod-resources-card__ram-table {
-    align-self: start;
-    margin-bottom: 0;
-
-    th,
-    td {
-      &:nth-child(1) {
-        padding-left: 0;
-        width: 50%;
-      }
-
-      &:nth-child(2) {
-        width: 6.25rem;
-      }
-
-      &:nth-child(3) {
-        padding-right: 0;
-        width: 50%;
-      }
-    }
-
-    td {
-      padding-bottom: $spv-inner--small;
-      padding-top: calc(#{$spv-inner--small} - 1px);
-    }
-
-    @media only screen and (min-width: $breakpoint-medium) {
-      margin-top: -$spv-inner--small;
-    }
-  }
-
-  .pod-resources-card__ram-chart {
-    display: block;
-    margin: 0 auto;
-  }
-
-  .pod-resources-card__vms-button {
-    line-height: 1;
-    margin: 0;
-    padding-left: $sph-inner--small * 0.5;
-    padding-right: $sph-inner--small;
-  }
-
-  // Base KVM resources card styling.
-  .pod-resources-card__section {
-    @extend %base-grid;
-    grid:
-      [row1-start] "header" min-content [row1-end]
-      [row2-start] "content" min-content [row2-end]
-      / 1fr;
-    grid-template-rows: min-content;
-
-    > :nth-child(1) {
-      grid-area: header;
-    }
-
-    > :nth-child(2) {
-      grid-area: content;
-    }
-
-    &:not(:first-of-type) {
-      padding-top: $spv-inner--large;
-      position: relative;
-
-      &::after {
-        background-color: $color-mid-light;
-        content: "";
-        height: 1px;
-        left: 0;
-        position: absolute;
-        right: 0;
-        top: $spv-inner--small;
-      }
-    }
-
-    @media only screen and (min-width: $breakpoint-medium) {
-      .pod-resources-card--wide & {
-        grid:
-          [row1-start] "header content content content" min-content [row1-end]
-          / minmax($donut-size, 1fr) 1fr 1fr 1fr;
-      }
-    }
-
-    @media only screen and (min-width: $breakpoint-medium) and (max-width: $breakpoint-x-large - 1px) {
-      grid:
-        [row1-start] "header content content content" min-content [row1-end]
-        / minmax($donut-size, 1fr) 1fr 1fr 1fr;
-    }
-  }
-
-  // Overwrites for aggregated view at large breakpoints
-  .pod-resources-card--aggregate {
-    padding-top: $spv-inner--small;
-
-    .p-card__content {
-      @extend %base-grid;
-      display: block;
-
-      @media only screen and (min-width: $breakpoint-large) {
-        display: grid;
-        grid-template-areas: "ram ram ram ram ram cpu cpu cpu vfs vfs vfs vms";
-        grid-template-columns: repeat(12, minmax(0, 1fr));
-        grid-template-rows: min-content;
-
-        .pod-resources-card__section {
-          padding: 0;
-
-          &:not(:first-of-type)::after {
-            bottom: -$spv-inner--large;
-            height: inherit;
-            left: -$sph-inner;
-            right: inherit;
-            top: -$spv-inner--small;
-            width: 1px;
-          }
-        }
-
-        .pod-resources-card__ram {
-          grid-area: ram;
-          grid-column-gap: $sph-inner;
-          grid-template-areas: "header content";
-          grid-template-columns: minmax($donut-size, 1fr) 3fr;
-        }
-
-        .pod-resources-card__cpu {
-          grid-area: cpu;
-          grid-template-areas:
-            "header header header"
-            "content content content";
-          grid-template-columns: repeat(3, minmax(0, 1fr));
-        }
-
-        .pod-resources-card__vfs {
-          grid-area: vfs;
-          grid-template-areas:
-            "header header header"
-            "content content content";
-          grid-template-columns: repeat(3, minmax(0, 1fr));
-        }
-
-        .pod-resources-card__vms {
-          grid-area: vms;
-          grid-template-areas:
-            "header"
-            "content";
-          grid-template-columns: 1fr;
-          text-align: right;
-        }
-      }
-
-      @media only screen and (min-width: $breakpoint-x-large) {
-        grid-template-areas: "ram ram ram ram cpu cpu cpu vfs vfs vfs vms vms";
-
-        .pod-resources-card__vms {
-          grid-template-areas: "header content";
-          grid-template-columns: repeat(2, minmax(0, 1fr));
-          text-align: left;
-        }
-      }
-    }
-  }
-}
