--- ui/src/app/kvm/components/PodResourcesCard/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./PodResourcesCard";
-export type { Props as PodResourcesCardProps } from "./PodResourcesCard";
