--- ui/src/app/kvm/components/PodStorage/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./PodStorage";
