--- ui/src/app/kvm/views/KVMDetails/KVMResources/KVMNumaResources/_index.scss
@@ -1,16 +1,0 @@
-@mixin KVMNumaResources {
-  .numa-resources-grid {
-    @extend %base-grid;
-    grid-template-columns: 1fr;
-
-    @media only screen and (min-width: $breakpoint-large) {
-      grid-template-columns: 1fr 1fr;
-    }
-
-    @media only screen and (min-width: $breakpoint-x-large) {
-      &:not(.is-wide) {
-        grid-template-columns: 1fr 1fr 1fr 1fr;
-      }
-    }
-  }
-}
