--- ui/src/app/kvm/views/KVMDetails/KVMResources/KVMNumaResources/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./KVMNumaResources";
