+++ ui/src/app/kvm/views/KVMDetails/KVMResources/KVMStorageCards/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./KVMStorageCards";
