+++ ui/src/app/kvm/views/KVMDetails/KVMResources/NumaResources/NumaResources.test.tsx
