+++ ui/src/app/kvm/views/KVMDetails/KVMResources/NumaResources/NumaResourcesCard/NumaResourcesCard.test.tsx
