+++ ui/src/app/kvm/views/KVMDetails/KVMResources/NumaResources/NumaResourcesCard/_index.scss
