+++ ui/src/app/kvm/views/KVMDetails/KVMResources/NumaResources/NumaResourcesCard/index.ts
