+++ ui/src/app/kvm/views/KVMDetails/KVMResources/NumaResources/_index.scss
