+++ ui/src/app/kvm/views/KVMDetails/KVMResources/NumaResources/index.ts
