+++ ui/src/app/kvm/views/KVMDetails/KVMResources/OverallResourcesCard/OverallCores/OverallCores.tsx
