+++ ui/src/app/kvm/views/KVMDetails/KVMResources/OverallResourcesCard/OverallCores/_index.scss
