+++ ui/src/app/kvm/views/KVMDetails/KVMResources/OverallResourcesCard/OverallRam/OverallRam.test.tsx
