+++ ui/src/app/kvm/views/KVMDetails/KVMResources/OverallResourcesCard/OverallRam/_index.scss
