+++ ui/src/app/kvm/views/KVMDetails/KVMResources/OverallResourcesCard/OverallResourcesCard.test.tsx
