+++ ui/src/app/kvm/views/KVMDetails/KVMResources/OverallResourcesCard/OverallResourcesCard.tsx
