+++ ui/src/app/kvm/views/KVMDetails/KVMResources/OverallResourcesCard/_index.scss
