+++ ui/src/app/kvm/views/KVMDetails/KVMResources/OverallResourcesCard/index.ts
