+++ ui/src/app/kvm/views/KVMDetails/KVMResources/ProjectResourcesCard/ProjectResourcesCard.tsx
