+++ ui/src/app/kvm/views/KVMDetails/KVMResources/ProjectResourcesCard/_index.scss
