+++ ui/src/app/kvm/views/KVMDetails/KVMResources/ProjectResourcesCard/index.ts
