--- ui/src/app/kvm/views/KVMDetails/LxdResources/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./LxdResources";
