+++ ui/src/scss/index.scss
@@ -93,20 +93,18 @@
 @import "~app/kvm/components/KVMActionFormWrapper/ComposeForm/StorageTable";
 @import "~app/kvm/components/KVMActionFormWrapper/ComposeForm/StorageTable/PoolSelect";
 @import "~app/kvm/components/PodMeter";
-@import "~app/kvm/components/PodResourcesCard";
-@import "~app/kvm/components/PodStorage";
 @import "~app/kvm/views/KVMDetails/CoreResources";
 @import "~app/kvm/views/KVMDetails/KVMDetailsHeader";
-@import "~app/kvm/views/KVMDetails/KVMResources/KVMNumaResources";
+@import "~app/kvm/views/KVMDetails/KVMResources/KVMStorageCards";
+@import "~app/kvm/views/KVMDetails/KVMResources/NumaResources";
+@import "~app/kvm/views/KVMDetails/KVMResources/NumaResources/NumaResourcesCard";
+@import "~app/kvm/views/KVMDetails/KVMResources/OverallResourcesCard";
+@import "~app/kvm/views/KVMDetails/KVMResources/ProjectResourcesCard";
+@import "~app/kvm/views/KVMDetails/KVMResources/OverallResourcesCard/OverallCores";
+@import "~app/kvm/views/KVMDetails/KVMResources/OverallResourcesCard/OverallRam";
 @import "~app/kvm/views/KVMDetails/LxdProject/ProjectSummaryCard";
 @import "~app/kvm/views/KVMDetails/LxdProject/ProjectVMs/VMsActionBar";
 @import "~app/kvm/views/KVMDetails/LxdProject/ProjectVMs/VMsTable";
-@import "~app/kvm/views/KVMDetails/LxdResources/NumaResources";
-@import "~app/kvm/views/KVMDetails/LxdResources/NumaResources/NumaResourcesCard";
-@import "~app/kvm/views/KVMDetails/LxdResources/OverallResourcesCard";
-@import "~app/kvm/views/KVMDetails/LxdResources/ProjectResourcesCard";
-@import "~app/kvm/views/KVMDetails/LxdResources/OverallResourcesCard/OverallCores";
-@import "~app/kvm/views/KVMDetails/LxdResources/OverallResourcesCard/OverallRam";
 @import "~app/kvm/views/KVMDetails/RamResources";
 @import "~app/kvm/views/KVMDetails/StorageResources";
 @import "~app/kvm/views/KVMDetails/VfResources";
@@ -121,8 +119,8 @@
 @include KVMComposeInterfacesTable;
 @include KVMComposeStorageTable;
 @include KVMDetailsHeader;
-@include KVMNumaResources;
 @include KVMPoolSelect;
+@include KVMStorageCards;
 @include KVMSubnetSelect;
 @include LxdTable;
 @include NumaResources;
@@ -131,8 +129,6 @@
 @include OverallRam;
 @include OverallResourcesCard;
 @include PodMeter;
-@include PodResourcesCard;
-@include PodStorage;
 @include ProjectResourcesCard;
 @include ProjectSummaryCard;
 @include RamResources;
